[set work attributes] Fix accidental key removal
The way we get edit data (through a sourceData block in the edit page source code) does not have information about work attributes, e.g. the key (tonality).

We keep the possibility to change the key even if it is already set in order to have the right data to post. Sending no data about the key removes the key from the work if it exists